deprecate formatify_ in favour of getDisplayValues
resolves legalese-io/legalese-io.github.io#94 and legalese-io/legalese-io.github.io#82

wherever possible we obtain the formatted value from getDisplayValues()

but in some cases we still have to call formatify_ anyway, like when a
value is synthesized rather than read from the spreadsheet.

the major change here is that % doesn't need to be placed in the
template, but appears as part of the value.

So, in future, if you see repeated %% signs, you should delete the %
from the template.