major bugfix: entitiesByName were not returned
when a spreadsheet has ~Include Entities~ the EntitesByName due to the
included sheet should bubble back up to the caller. in the past, it
didn't. now, it does.

how did anything ever work before? this is a schrödinbug.
http://programmers.stackexchange.com/questions/96973/whats-a-schr%C3%B6dinbug

this may now uncover all sorts of other weirdness.

if only we had a proper test suite.